@@ -1,14 +1,17 @@
 import React from 'react';
+import { Provider } from 'react-redux';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
+import store from '../src/components/store'
 
 ReactDOM.render(
+  <Provider store = {store}>
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root')
+  </Provider>,document.getElementById('root')
 );
 
 // If you want your app to work offline and load faster, you can change
